+++ asciinema/__main__.py
@@ -8,6 +8,7 @@ import asciinema.config as config
 from asciinema.commands.auth import AuthCommand
 from asciinema.commands.record import RecordCommand
 from asciinema.commands.play import PlayCommand
+from asciinema.commands.cat import CatCommand
 from asciinema.commands.upload import UploadCommand
 from asciinema.api import Api
 
@@ -21,21 +22,26 @@ def positive_float(value):
 
 
 def rec_command(args, config):
-    api = Api(config.api_url, os.environ.get("USER"), config.api_token)
-    return RecordCommand(api, args.filename, args.command, args.title, args.yes, args.quiet, args.max_wait)
+    api = Api(config.api_url, os.environ.get("USER"), config.install_id)
+    return RecordCommand(api, args)
 
 
 def play_command(args, config):
-    return PlayCommand(args.filename, args.max_wait, args.speed)
+    return PlayCommand(args.filename, args.idle_time_limit, args.speed)
+
+
+def cat_command(args, config):
+    return CatCommand(args.filename)
 
 
 def upload_command(args, config):
-    api = Api(config.api_url, os.environ.get("USER"), config.api_token)
+    api = Api(config.api_url, os.environ.get("USER"), config.install_id)
     return UploadCommand(api, args.filename)
 
 
 def auth_command(args, config):
-    return AuthCommand(config.api_url, config.api_token)
+    api = Api(config.api_url, os.environ.get("USER"), config.install_id)
+    return AuthCommand(api)
 
 
 def maybe_str(v):
@@ -48,7 +54,11 @@ def main():
         print("asciinema needs a UTF-8 native locale to run. Check the output of `locale` command.")
         sys.exit(1)
 
-    cfg = config.load()
+    try:
+        cfg = config.load()
+    except config.ConfigError as e:
+        sys.stderr.write(str(e) + '\n')
+        sys.exit(1)
 
     # create the top-level parser
     parser = argparse.ArgumentParser(
@@ -57,15 +67,17 @@ def main():
   Record terminal and upload it to asciinema.org:
     \x1b[1masciinema rec\x1b[0m
   Record terminal to local file:
-    \x1b[1masciinema rec demo.json\x1b[0m
+    \x1b[1masciinema rec demo.cast\x1b[0m
   Record terminal and upload it to asciinema.org, specifying title:
     \x1b[1masciinema rec -t "My git tutorial"\x1b[0m
-  Record terminal to local file, "trimming" longer pauses to max 2.5 sec:
-    \x1b[1masciinema rec -w 2.5 demo.json\x1b[0m
+  Record terminal to local file, limiting idle time to max 2.5 sec:
+    \x1b[1masciinema rec -i 2.5 demo.cast\x1b[0m
   Replay terminal recording from local file:
-    \x1b[1masciinema play demo.json\x1b[0m
+    \x1b[1masciinema play demo.cast\x1b[0m
   Replay terminal recording hosted on asciinema.org:
     \x1b[1masciinema play https://asciinema.org/a/difqlgx86ym6emrmd8u62yqu8\x1b[0m
+  Print full output of recorded session:
+    \x1b[1masciinema cat demo.cast\x1b[0m
 
 For help on a specific command run:
   \x1b[1masciinema <command> -h\x1b[0m""",
@@ -77,9 +89,14 @@ For help on a specific command run:
 
     # create the parser for the "rec" command
     parser_rec = subparsers.add_parser('rec', help='Record terminal session')
+    parser_rec.add_argument('--stdin', help='enable stdin recording, disabled by default', action='store_true', default=cfg.record_stdin)
+    parser_rec.add_argument('--append', help='append to existing recording', action='store_true', default=False)
+    parser_rec.add_argument('--raw', help='save only raw stdout output', action='store_true', default=False)
+    parser_rec.add_argument('--overwrite', help='overwrite the file if it already exists', action='store_true', default=False)
     parser_rec.add_argument('-c', '--command', help='command to record, defaults to $SHELL', default=cfg.record_command)
+    parser_rec.add_argument('-e', '--env', help='list of environment variables to capture, defaults to ' + config.DEFAULT_RECORD_ENV, default=cfg.record_env)
     parser_rec.add_argument('-t', '--title', help='title of the asciicast')
-    parser_rec.add_argument('-w', '--max-wait', help='limit recorded terminal inactivity to max <sec> seconds (can be fractional)', type=positive_float, default=maybe_str(cfg.record_max_wait))
+    parser_rec.add_argument('-i', '--idle-time-limit', help='limit recorded idle time to given number of seconds', type=positive_float, default=maybe_str(cfg.record_idle_time_limit))
     parser_rec.add_argument('-y', '--yes', help='answer "yes" to all prompts (e.g. upload confirmation)', action='store_true', default=cfg.record_yes)
     parser_rec.add_argument('-q', '--quiet', help='be quiet, suppress all notices/warnings (implies -y)', action='store_true', default=cfg.record_quiet)
     parser_rec.add_argument('filename', nargs='?', default='', help='filename/path to save the recording to')
@@ -87,10 +104,15 @@ For help on a specific command run:
 
     # create the parser for the "play" command
     parser_play = subparsers.add_parser('play', help='Replay terminal session')
-    parser_play.add_argument('-w', '--max-wait', help='limit terminal inactivity to max <sec> seconds (can be fractional)', type=positive_float, default=maybe_str(cfg.play_max_wait))
+    parser_play.add_argument('-i', '--idle-time-limit', help='limit idle time during playback to given number of seconds', type=positive_float, default=maybe_str(cfg.play_idle_time_limit))
     parser_play.add_argument('-s', '--speed', help='playback speedup (can be fractional)', type=positive_float, default=cfg.play_speed)
     parser_play.add_argument('filename', help='local path, http/ipfs URL or "-" (read from stdin)')
     parser_play.set_defaults(func=play_command)
+
+    # create the parser for the "cat" command
+    parser_cat = subparsers.add_parser('cat', help='Print full output of terminal session')
+    parser_cat.add_argument('filename', help='local path, http/ipfs URL or "-" (read from stdin)')
+    parser_cat.set_defaults(func=cat_command)
 
     # create the parser for the "upload" command
     parser_upload = subparsers.add_parser('upload', help='Upload locally saved terminal session to asciinema.org')