+++ asciinema/api.py
@@ -1,5 +1,6 @@
 import platform
 import re
+from urllib.parse import urlparse
 
 from asciinema import __version__
 from asciinema.urllib_http_adapter import URLLibHttpAdapter
@@ -12,14 +13,17 @@ class APIError(Exception):
 
 class Api:
 
-    def __init__(self, url, user, token, http_adapter=None):
+    def __init__(self, url, user, install_id, http_adapter=None):
         self.url = url
         self.user = user
-        self.token = token
+        self.install_id = install_id
         self.http_adapter = http_adapter if http_adapter is not None else URLLibHttpAdapter()
 
+    def hostname(self):
+        return urlparse(self.url).hostname
+
     def auth_url(self):
-        return "{}/connect/{}".format(self.url, self.token)
+        return "{}/connect/{}".format(self.url, self.install_id)
 
     def upload_url(self):
         return "{}/api/asciicasts".format(self.url)
@@ -29,10 +33,10 @@ class Api:
             try:
                 status, headers, body = self.http_adapter.post(
                     self.upload_url(),
-                    files={"asciicast": ("asciicast.json", f)},
+                    files={"asciicast": ("ascii.cast", f)},
                     headers=self._headers(),
                     username=self.user,
-                    password=self.token
+                    password=self.install_id
                 )
             except HTTPConnectionError as e:
                 raise APIError(str(e))
@@ -57,7 +61,7 @@ class Api:
     def _handle_error(self, status, body):
         errors = {
             400: "Invalid request: %s" % body,
-            401: "Invalid or revoked recorder token",
+            401: "Invalid or revoked install ID",
             404: "API endpoint not found. This asciinema version may no longer be supported. Please upgrade to the latest version.",
             413: "Sorry, your asciicast is too big.",
             422: "Invalid asciicast: %s" % body,