+++ tests/pty_recorder_test.py
@@ -4,7 +4,6 @@ import pty
 from nose.tools import assert_equal
 from .test_helper import Test
 
-from asciinema.stdout import Stdout
 from asciinema.pty_recorder import PtyRecorder
 
 
@@ -12,13 +11,12 @@ class FakeStdout:
 
     def __init__(self):
         self.data = []
-        self.closed = False
 
-    def write(self, data):
+    def write_stdout(self, data):
         self.data.append(data)
 
-    def close(self):
-        self.close = True
+    def write_stdin(self, data):
+        pass
 
 
 class TestPtyRecorder(Test):